Revert "[be-012] added populateDatabase task for Gradle"
[BE-P-016] minor fixes after merge and added check that imageUrl contains bucketUrl for image to be deleted

+++ src/main/java/com/spd/baraholka/image/service/ImageServiceImpl.java
@@ -139,7 +139,11 @@ public class ImageServiceImpl implements ImageService {
 
     @Override
     public void deleteImage(String imageUrl) {
-        aws3Service.deleteImageFromS3Bucket(imageUrl);
+        Objects.requireNonNull(imageUrl);
+        String bucketUrl = aws3Service.getAmazonDomain() + aws3Service.getBucketName();
+        if (imageUrl.contains(bucketUrl)) {
+            aws3Service.deleteImageFromS3Bucket(imageUrl);
+        }
     }
 
     private String generateFileName(ImageResource imageResource) {
